# src/main/resources/static/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Swagger SPA - OpenAPI 3.0
  contact:
    email: ismailrzouki7@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:8080/api/v1
   
paths:
  /greeting:
    get:
      tags: 
        - Greeting
      operationId: greeting
      summary: Hello world openApi
      responses:
        '200':
          description: Greeting from SPA.
          content:
            text/plain:
              schema:
                type: string
                example: Hallo Developer!
  /user/auth:
    post:
      tags:
        - Authentication
      operationId: auth
      summary: User Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CredentialPayload'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: json
  
  /user/create:
    post:
      tags:
        - Authentication
      operationId: create
      summary: User Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPayload'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: json

  /product/delete:
    post:
      tags:
        - Product Managment
      operationId: deleteProduct
      summary: Delete Product by docId
      parameters:
        - in: query
          name: docId
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: json
                
  /product/main:
    get:
      tags:
        - Product Managment
      operationId: mainProduct
      summary: Retrieve the main product for the app
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'

  /product/add:
    post:
      tags:
        - Product Managment
      operationId: addProduct
      summary: add Product to firestore.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: json
                
  /payment/create:
    post:
      tags:
        - Payment Gateway
      operationId: createPI
      summary: create paymentIntent
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentPayload'
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: payment intent id
       
                
  /payment/customer:
    post:
      tags:
        - Payment Gateway
      operationId: addCustomer
      summary: add stripe customer.
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: email
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'

  /payment/confirm:
    post:
      tags:
        - Payment Gateway
      operationId: confirmPI
      summary: confirm paymentIntent
      parameters:
        - in: query
          name: paymentIntent
          schema:
            type: string
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
                example: paymentIntent ID.
                
components:
  schemas:
  
    UserDto:
      type: object
      properties:
        email:
          type: string
          example: "ismaa@ecommapp.com"
        password:
          type: string
          example: "plainText"
        returnSecureToken:
          type: boolean
          example: true
          
    CredentialPayload:
      type: object
      properties:
        email:
          type: string
          example: "ismaa@ecommapp.com"
        password:
          type: string
          example: "plainText"
        returnSecureToken:
          type: boolean
          example: true
    
    RegisterPayload:
      type: object
      properties:
        email:
          type: string
          example: "ismaa@ecommapp.com"
        password:
          type: string
          example: "plainText"
        phone:
          type: string
          example: +212654741162
        photo:
          type: string
          example: https://i.imgur.com/RzikSN4.png
    
    ProductDto:
      type: object
      properties:
        id:
          type: string
          example: eafa2FE4FAA334
        name:
          type: string
          example: "KEto"
        price:
          type: string
          example: "120"
        description:
          type: string
          example: Hairlost Keto Bottle for regen Damaged hair or lost, Discover our handmade/bio oil for your hair.
        isActive:
          type: boolean
          example: true
        isDeleted:
          type: boolean
          example: false
        isPrincipal:
          type: boolean
          example: true
        mainImage:
          type: string
          example: https://i.imgur.com/RzikSN4.png
        subImages:
          type: array
          items:
            type: string

    Customer:
      type: object
      properties:
        name: 
          $ref: 'Customer'
     
    PaymentPayload:
      type: object
      properties:
        amount:
          type: string
          example: 299.99
        currenrcy:
          type: string
          example: "usd"
        description:
          type: string
          example: desc
        customer:
          type: string
          example: ismaa
        automaticPayment:
           type: boolean
           example: true